third_party/WebKit/LayoutTests/media/W3C/video/currentSrc/currentSrc_empty_if_no_src-expected.txt

video.currentSrc - empty if no source

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS "" is ""

TEST COMPLETE
PASS successfullyParsed is true

TEST COMPLETE
spec reference
